# app/controllers.go
// Code generated by goagen v1.1.0-dirty, command line:
// $ goagen
// --design=github.com/tikasan/eventory/design
// --out=$(GOPATH)
// --version=v1.1.0-dirty
//
// API "eventory": Application Controllers
//
// The content of this file is auto-generated, DO NOT MODIFY

package app

import (
	"github.com/goadesign/goa"
	"github.com/goadesign/goa/cors"
	"golang.org/x/net/context"
	"net/http"
)

// initService sets up the service encoders, decoders and mux.
func initService(service *goa.Service) {
	// Setup encoders and decoders
	service.Encoder.Register(goa.NewJSONEncoder, "application/json")
	service.Encoder.Register(goa.NewGobEncoder, "application/gob", "application/x-gob")
	service.Encoder.Register(goa.NewXMLEncoder, "application/xml")
	service.Decoder.Register(goa.NewJSONDecoder, "application/json")
	service.Decoder.Register(goa.NewGobDecoder, "application/gob", "application/x-gob")
	service.Decoder.Register(goa.NewXMLDecoder, "application/xml")

	// Setup default encoder and decoder
	service.Encoder.Register(goa.NewJSONEncoder, "*/*")
	service.Decoder.Register(goa.NewJSONDecoder, "*/*")
}

// CronController is the controller interface for the Cron actions.
type CronController interface {
	goa.Muxer
	AppendGenre(*AppendGenreCronContext) error
	FixUserFollow(*FixUserFollowCronContext) error
	NewEventFetch(*NewEventFetchCronContext) error
}

// MountCronController "mounts" a Cron resource controller on the given service.
func MountCronController(service *goa.Service, ctrl CronController) {
	initService(service)
	var h goa.Handler
	service.Mux.Handle("OPTIONS", "/api/v2/cron/events/appendgenre", ctrl.MuxHandler("preflight", handleCronOrigin(cors.HandlePreflight()), nil))
	service.Mux.Handle("OPTIONS", "/api/v2/cron/user/events/fixfollow", ctrl.MuxHandler("preflight", handleCronOrigin(cors.HandlePreflight()), nil))
	service.Mux.Handle("OPTIONS", "/api/v2/cron/events/fetch", ctrl.MuxHandler("preflight", handleCronOrigin(cors.HandlePreflight()), nil))

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewAppendGenreCronContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.AppendGenre(rctx)
	}
	h = handleSecurity("cronToken", h)
	h = handleCronOrigin(h)
	service.Mux.Handle("GET", "/api/v2/cron/events/appendgenre", ctrl.MuxHandler("AppendGenre", h, nil))
	service.LogInfo("mount", "ctrl", "Cron", "action", "AppendGenre", "route", "GET /api/v2/cron/events/appendgenre", "security", "cronToken")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewFixUserFollowCronContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.FixUserFollow(rctx)
	}
	h = handleSecurity("cronToken", h)
	h = handleCronOrigin(h)
	service.Mux.Handle("GET", "/api/v2/cron/user/events/fixfollow", ctrl.MuxHandler("FixUserFollow", h, nil))
	service.LogInfo("mount", "ctrl", "Cron", "action", "FixUserFollow", "route", "GET /api/v2/cron/user/events/fixfollow", "security", "cronToken")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewNewEventFetchCronContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.NewEventFetch(rctx)
	}
	h = handleSecurity("cronToken", h)
	h = handleCronOrigin(h)
	service.Mux.Handle("GET", "/api/v2/cron/events/fetch", ctrl.MuxHandler("NewEventFetch", h, nil))
	service.LogInfo("mount", "ctrl", "Cron", "action", "NewEventFetch", "route", "GET /api/v2/cron/events/fetch", "security", "cronToken")
}

// handleCronOrigin applies the CORS response headers corresponding to the origin.
func handleCronOrigin(h goa.Handler) goa.Handler {

	return func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		origin := req.Header.Get("Origin")
		if origin == "" {
			// Not a CORS request
			return h(ctx, rw, req)
		}
		if cors.MatchOrigin(origin, "*") {
			ctx = goa.WithLogContext(ctx, "origin", origin)
			rw.Header().Set("Access-Control-Allow-Origin", origin)
			rw.Header().Set("Access-Control-Max-Age", "600")
			rw.Header().Set("Access-Control-Allow-Credentials", "true")
			if acrm := req.Header.Get("Access-Control-Request-Method"); acrm != "" {
				// We are handling a preflight request
				rw.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT")
			}
			return h(ctx, rw, req)
		}

		return h(ctx, rw, req)
	}
}

// EventsController is the controller interface for the Events actions.
type EventsController interface {
	goa.Muxer
	Keep(*KeepEventsContext) error
	List(*ListEventsContext) error
}

// MountEventsController "mounts" a Events resource controller on the given service.
func MountEventsController(service *goa.Service, ctrl EventsController) {
	initService(service)
	var h goa.Handler
	service.Mux.Handle("OPTIONS", "/api/v2/events/:eventID/keep", ctrl.MuxHandler("preflight", handleEventsOrigin(cors.HandlePreflight()), nil))
	service.Mux.Handle("OPTIONS", "/api/v2/events/genre/:id", ctrl.MuxHandler("preflight", handleEventsOrigin(cors.HandlePreflight()), nil))
	service.Mux.Handle("OPTIONS", "/api/v2/events/new", ctrl.MuxHandler("preflight", handleEventsOrigin(cors.HandlePreflight()), nil))
	service.Mux.Handle("OPTIONS", "/api/v2/events/keep", ctrl.MuxHandler("preflight", handleEventsOrigin(cors.HandlePreflight()), nil))
	service.Mux.Handle("OPTIONS", "/api/v2/events/nokeep", ctrl.MuxHandler("preflight", handleEventsOrigin(cors.HandlePreflight()), nil))
	service.Mux.Handle("OPTIONS", "/api/v2/events/popular", ctrl.MuxHandler("preflight", handleEventsOrigin(cors.HandlePreflight()), nil))
	service.Mux.Handle("OPTIONS", "/api/v2/events/recommend", ctrl.MuxHandler("preflight", handleEventsOrigin(cors.HandlePreflight()), nil))

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewKeepEventsContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Keep(rctx)
	}
	h = handleSecurity("userToken", h)
	h = handleEventsOrigin(h)
	service.Mux.Handle("PUT", "/api/v2/events/:eventID/keep", ctrl.MuxHandler("Keep", h, nil))
	service.LogInfo("mount", "ctrl", "Events", "action", "Keep", "route", "PUT /api/v2/events/:eventID/keep", "security", "userToken")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewListEventsContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.List(rctx)
	}
	h = handleSecurity("userToken", h)
	h = handleEventsOrigin(h)
	service.Mux.Handle("GET", "/api/v2/events/genre/:id", ctrl.MuxHandler("List", h, nil))
	service.LogInfo("mount", "ctrl", "Events", "action", "List", "route", "GET /api/v2/events/genre/:id", "security", "userToken")
	service.Mux.Handle("GET", "/api/v2/events/new", ctrl.MuxHandler("List", h, nil))
	service.LogInfo("mount", "ctrl", "Events", "action", "List", "route", "GET /api/v2/events/new", "security", "userToken")
	service.Mux.Handle("GET", "/api/v2/events/keep", ctrl.MuxHandler("List", h, nil))
	service.LogInfo("mount", "ctrl", "Events", "action", "List", "route", "GET /api/v2/events/keep", "security", "userToken")
	service.Mux.Handle("GET", "/api/v2/events/nokeep", ctrl.MuxHandler("List", h, nil))
	service.LogInfo("mount", "ctrl", "Events", "action", "List", "route", "GET /api/v2/events/nokeep", "security", "userToken")
	service.Mux.Handle("GET", "/api/v2/events/popular", ctrl.MuxHandler("List", h, nil))
	service.LogInfo("mount", "ctrl", "Events", "action", "List", "route", "GET /api/v2/events/popular", "security", "userToken")
	service.Mux.Handle("GET", "/api/v2/events/recommend", ctrl.MuxHandler("List", h, nil))
	service.LogInfo("mount", "ctrl", "Events", "action", "List", "route", "GET /api/v2/events/recommend", "security", "userToken")
}

// handleEventsOrigin applies the CORS response headers corresponding to the origin.
func handleEventsOrigin(h goa.Handler) goa.Handler {

	return func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		origin := req.Header.Get("Origin")
		if origin == "" {
			// Not a CORS request
			return h(ctx, rw, req)
		}
		if cors.MatchOrigin(origin, "*") {
			ctx = goa.WithLogContext(ctx, "origin", origin)
			rw.Header().Set("Access-Control-Allow-Origin", origin)
			rw.Header().Set("Access-Control-Max-Age", "600")
			rw.Header().Set("Access-Control-Allow-Credentials", "true")
			if acrm := req.Header.Get("Access-Control-Request-Method"); acrm != "" {
				// We are handling a preflight request
				rw.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT")
			}
			return h(ctx, rw, req)
		}

		return h(ctx, rw, req)
	}
}

// GenresController is the controller interface for the Genres actions.
type GenresController interface {
	goa.Muxer
	Create(*CreateGenresContext) error
	Follow(*FollowGenresContext) error
	List(*ListGenresContext) error
}

// MountGenresController "mounts" a Genres resource controller on the given service.
func MountGenresController(service *goa.Service, ctrl GenresController) {
	initService(service)
	var h goa.Handler
	service.Mux.Handle("OPTIONS", "/api/v2/genres/new", ctrl.MuxHandler("preflight", handleGenresOrigin(cors.HandlePreflight()), nil))
	service.Mux.Handle("OPTIONS", "/api/v2/genres/:genreID/follow", ctrl.MuxHandler("preflight", handleGenresOrigin(cors.HandlePreflight()), nil))
	service.Mux.Handle("OPTIONS", "/api/v2/genres", ctrl.MuxHandler("preflight", handleGenresOrigin(cors.HandlePreflight()), nil))

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewCreateGenresContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Create(rctx)
	}
	h = handleSecurity("userToken", h)
	h = handleGenresOrigin(h)
	service.Mux.Handle("POST", "/api/v2/genres/new", ctrl.MuxHandler("Create", h, nil))
	service.LogInfo("mount", "ctrl", "Genres", "action", "Create", "route", "POST /api/v2/genres/new", "security", "userToken")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewFollowGenresContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Follow(rctx)
	}
	h = handleSecurity("userToken", h)
	h = handleGenresOrigin(h)
	service.Mux.Handle("PUT", "/api/v2/genres/:genreID/follow", ctrl.MuxHandler("Follow", h, nil))
	service.LogInfo("mount", "ctrl", "Genres", "action", "Follow", "route", "PUT /api/v2/genres/:genreID/follow", "security", "userToken")
	service.Mux.Handle("DELETE", "/api/v2/genres/:genreID/follow", ctrl.MuxHandler("Follow", h, nil))
	service.LogInfo("mount", "ctrl", "Genres", "action", "Follow", "route", "DELETE /api/v2/genres/:genreID/follow", "security", "userToken")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewListGenresContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.List(rctx)
	}
	h = handleSecurity("userToken", h)
	h = handleGenresOrigin(h)
	service.Mux.Handle("GET", "/api/v2/genres", ctrl.MuxHandler("List", h, nil))
	service.LogInfo("mount", "ctrl", "Genres", "action", "List", "route", "GET /api/v2/genres", "security", "userToken")
}

// handleGenresOrigin applies the CORS response headers corresponding to the origin.
func handleGenresOrigin(h goa.Handler) goa.Handler {

	return func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		origin := req.Header.Get("Origin")
		if origin == "" {
			// Not a CORS request
			return h(ctx, rw, req)
		}
		if cors.MatchOrigin(origin, "*") {
			ctx = goa.WithLogContext(ctx, "origin", origin)
			rw.Header().Set("Access-Control-Allow-Origin", origin)
			rw.Header().Set("Access-Control-Max-Age", "600")
			rw.Header().Set("Access-Control-Allow-Credentials", "true")
			if acrm := req.Header.Get("Access-Control-Request-Method"); acrm != "" {
				// We are handling a preflight request
				rw.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT")
			}
			return h(ctx, rw, req)
		}

		return h(ctx, rw, req)
	}
}

// PrefsController is the controller interface for the Prefs actions.
type PrefsController interface {
	goa.Muxer
	Follow(*FollowPrefsContext) error
}

// MountPrefsController "mounts" a Prefs resource controller on the given service.
func MountPrefsController(service *goa.Service, ctrl PrefsController) {
	initService(service)
	var h goa.Handler
	service.Mux.Handle("OPTIONS", "/api/v2/prefs/:prefID/follow", ctrl.MuxHandler("preflight", handlePrefsOrigin(cors.HandlePreflight()), nil))

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewFollowPrefsContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Follow(rctx)
	}
	h = handleSecurity("userToken", h)
	h = handlePrefsOrigin(h)
	service.Mux.Handle("PUT", "/api/v2/prefs/:prefID/follow", ctrl.MuxHandler("Follow", h, nil))
	service.LogInfo("mount", "ctrl", "Prefs", "action", "Follow", "route", "PUT /api/v2/prefs/:prefID/follow", "security", "userToken")
	service.Mux.Handle("DELETE", "/api/v2/prefs/:prefID/follow", ctrl.MuxHandler("Follow", h, nil))
	service.LogInfo("mount", "ctrl", "Prefs", "action", "Follow", "route", "DELETE /api/v2/prefs/:prefID/follow", "security", "userToken")
}

// handlePrefsOrigin applies the CORS response headers corresponding to the origin.
func handlePrefsOrigin(h goa.Handler) goa.Handler {

	return func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		origin := req.Header.Get("Origin")
		if origin == "" {
			// Not a CORS request
			return h(ctx, rw, req)
		}
		if cors.MatchOrigin(origin, "*") {
			ctx = goa.WithLogContext(ctx, "origin", origin)
			rw.Header().Set("Access-Control-Allow-Origin", origin)
			rw.Header().Set("Access-Control-Max-Age", "600")
			rw.Header().Set("Access-Control-Allow-Credentials", "true")
			if acrm := req.Header.Get("Access-Control-Request-Method"); acrm != "" {
				// We are handling a preflight request
				rw.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT")
			}
			return h(ctx, rw, req)
		}

		return h(ctx, rw, req)
	}
}

// UsersController is the controller interface for the Users actions.
type UsersController interface {
	goa.Muxer
	Login(*LoginUsersContext) error
	RegularCreate(*RegularCreateUsersContext) error
	Status(*StatusUsersContext) error
	TmpCreate(*TmpCreateUsersContext) error
}

// MountUsersController "mounts" a Users resource controller on the given service.
func MountUsersController(service *goa.Service, ctrl UsersController) {
	initService(service)
	var h goa.Handler
	service.Mux.Handle("OPTIONS", "/api/v2/users/login", ctrl.MuxHandler("preflight", handleUsersOrigin(cors.HandlePreflight()), nil))
	service.Mux.Handle("OPTIONS", "/api/v2/users/new", ctrl.MuxHandler("preflight", handleUsersOrigin(cors.HandlePreflight()), nil))
	service.Mux.Handle("OPTIONS", "/api/v2/users/status", ctrl.MuxHandler("preflight", handleUsersOrigin(cors.HandlePreflight()), nil))
	service.Mux.Handle("OPTIONS", "/api/v2/users/tmp", ctrl.MuxHandler("preflight", handleUsersOrigin(cors.HandlePreflight()), nil))

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewLoginUsersContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Login(rctx)
	}
	h = handleSecurity("userToken", h)
	h = handleUsersOrigin(h)
	service.Mux.Handle("POST", "/api/v2/users/login", ctrl.MuxHandler("Login", h, nil))
	service.LogInfo("mount", "ctrl", "Users", "action", "Login", "route", "POST /api/v2/users/login", "security", "userToken")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewRegularCreateUsersContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.RegularCreate(rctx)
	}
	h = handleSecurity("userToken", h)
	h = handleUsersOrigin(h)
	service.Mux.Handle("POST", "/api/v2/users/new", ctrl.MuxHandler("RegularCreate", h, nil))
	service.LogInfo("mount", "ctrl", "Users", "action", "RegularCreate", "route", "POST /api/v2/users/new", "security", "userToken")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewStatusUsersContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Status(rctx)
	}
	h = handleSecurity("userToken", h)
	h = handleUsersOrigin(h)
	service.Mux.Handle("PUT", "/api/v2/users/status", ctrl.MuxHandler("Status", h, nil))
	service.LogInfo("mount", "ctrl", "Users", "action", "Status", "route", "PUT /api/v2/users/status", "security", "userToken")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewTmpCreateUsersContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.TmpCreate(rctx)
	}
	h = handleUsersOrigin(h)
	service.Mux.Handle("POST", "/api/v2/users/tmp", ctrl.MuxHandler("TmpCreate", h, nil))
	service.LogInfo("mount", "ctrl", "Users", "action", "TmpCreate", "route", "POST /api/v2/users/tmp")
}

// handleUsersOrigin applies the CORS response headers corresponding to the origin.
func handleUsersOrigin(h goa.Handler) goa.Handler {

	return func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		origin := req.Header.Get("Origin")
		if origin == "" {
			// Not a CORS request
			return h(ctx, rw, req)
		}
		if cors.MatchOrigin(origin, "*") {
			ctx = goa.WithLogContext(ctx, "origin", origin)
			rw.Header().Set("Access-Control-Allow-Origin", origin)
			rw.Header().Set("Access-Control-Max-Age", "600")
			rw.Header().Set("Access-Control-Allow-Credentials", "true")
			if acrm := req.Header.Get("Access-Control-Request-Method"); acrm != "" {
				// We are handling a preflight request
				rw.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT")
			}
			return h(ctx, rw, req)
		}

		return h(ctx, rw, req)
	}
}