# swagger/swagger.yaml
basePath: /api/v2
consumes:
- application/json
- application/xml
- application/gob
- application/x-gob
definitions:
...
  error:
    description: Error response media type (default view)
    example:
      code: invalid_value
      detail: Value of ID must be an integer
      id: 3F1FKVRR
      meta:
        timestamp: 1.458609e+09
      status: "400"
    properties:
      code:
        description: an application-specific error code, expressed as a string value.
        example: invalid_value
        type: string
      detail:
        description: a human-readable explanation specific to this occurrence of the
          problem.
        example: Value of ID must be an integer
        type: string
      id:
        description: a unique identifier for this particular occurrence of the problem.
        example: 3F1FKVRR
        type: string
      meta:
        additionalProperties: true
        description: a meta object containing non-standard meta-information about
          the error.
        example:
          timestamp: 1.458609e+09
        type: object
      status:
        description: the HTTP status code applicable to this problem, expressed as
          a string value.
        example: "400"
        type: string
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
...
host: localhost:8080
info:
  description: 'ITイベント情報集アプリAPI '
  license:
    name: MIT
    url: https://github.com/tikasan/eventory/blob/master/LICENSE
  title: 'eventory: ITイベント収集アプリ'
  version: ""
paths:
  /cron/events/appendgenre:
    get:
      description: <b>イベントにジャンルを付加する<b>
      operationId: cron#append genre
      produces:
      - application/vnd.goa.error
      - text/plain
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized
      schemes:
      - http
      security:
      - cronToken: []
      summary: append genre cron
      tags:
      - cron
  /cron/events/fetch:
    get:
      description: <b>最新イベント情報の取得<b>
      operationId: cron#new event fetch
      produces:
      - application/vnd.goa.error
      - text/plain
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized
      schemes:
      - http
      security:
      - cronToken: []
      summary: new event fetch cron
      tags:
      - cron
  /cron/user/events/fixfollow:
    get:
      description: "<b>イベントフォロー操作の確定</b><br>\n\t\tuser_follow_eventsテーブルのbatch_processedをtrueに変更する"
      operationId: cron#fix user follow
      produces:
      - application/vnd.goa.error
      - text/plain
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized
      schemes:
      - http
      security:
      - cronToken: []
      summary: fix user follow cron
      tags:
      - cron
  /events/{eventID}/keep:
    put:
      description: "<b>イベントお気に入り操作</b><br>\n\t\tisKeepがtrueだった場合はフォロー、falseの場合はアンフォローとする。<br>\n\t\t存在しないイベントへのリクエストは404エラーを返す。"
      operationId: events#keep
      parameters:
      - description: イベントID
        in: path
        name: eventID
        required: true
        type: integer
      - description: キープ操作
        in: query
        name: isKeep
        required: true
        type: boolean
      produces:
      - text/plain
      - application/vnd.goa.error
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      schemes:
      - http
      security:
      - userToken: []
      summary: keep events
      tags:
      - events
  /events/genre/{id}:
    get:
      description: "<b>イベント情報取得</b><br>\n\t\t<ul>\n\t\t\t<li>/genre/:id -> ジャンル別新着情報</li>\n\t\t\t<li>/new
        -> ユーザー別新着情報</li>\n\t\t\t<li>/keep -> ユーザーがキープしているイベント</li>\n\t\t\t<li>/nokeep
        -> ユーザーが興味なしにしたイベント</li>\n\t\t\t<li>/popular -> キープ数が多い。注目されているイベント</li>\n\t\t\t<li>/recommend
        -> ユーザー属性に合わせたおすすめイベント</li>\n\t\t</ul>\n\t\tイベントの情報は区切って送信され、スクロールイベントで次のページのイベント情報を取得することを想定している。<br>\n\t\tまた、キープや興味なしの操作は１日に１回行われるバッチ処理時に確定されるまでは、分類されずに表示される。"
      operationId: events#list
      parameters:
      - in: path
        name: id
        required: true
        type: string
      - default: 0
        description: ページ(1->2->3->4)
        in: query
        minimum: 1
        name: page
        required: false
        type: integer
      - default: ""
        description: キーワード検索
        in: query
        name: q
        required: false
        type: string
      - default: ""
        description: ソート
        enum:
        - created_asc
        - created_desc
        - ""
        in: query
        name: sort
        required: false
        type: string
      produces:
      - application/vnd.goa.error
      - application/vnd.event+json; type=collection
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/EventCollection'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized
      schemes:
      - http
      security:
      - userToken: []
      summary: list events
      tags:
      - events
  /events/keep:
    get:
      description: "<b>イベント情報取得</b><br>\n\t\t<ul>\n\t\t\t<li>/genre/:id -> ジャンル別新着情報</li>\n\t\t\t<li>/new
        -> ユーザー別新着情報</li>\n\t\t\t<li>/keep -> ユーザーがキープしているイベント</li>\n\t\t\t<li>/nokeep
        -> ユーザーが興味なしにしたイベント</li>\n\t\t\t<li>/popular -> キープ数が多い。注目されているイベント</li>\n\t\t\t<li>/recommend
        -> ユーザー属性に合わせたおすすめイベント</li>\n\t\t</ul>\n\t\tイベントの情報は区切って送信され、スクロールイベントで次のページのイベント情報を取得することを想定している。<br>\n\t\tまた、キープや興味なしの操作は１日に１回行われるバッチ処理時に確定されるまでは、分類されずに表示される。"
      operationId: events#list#2
      parameters:
      - in: query
        name: id
        required: false
        type: string
      - default: 0
        description: ページ(1->2->3->4)
        in: query
        minimum: 1
        name: page
        required: false
        type: integer
      - default: ""
        description: キーワード検索
        in: query
        name: q
        required: false
        type: string
      - default: ""
        description: ソート
        enum:
        - created_asc
        - created_desc
        - ""
        in: query
        name: sort
        required: false
        type: string
      produces:
      - application/vnd.goa.error
      - application/vnd.event+json; type=collection
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/EventCollection'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized
      schemes:
      - http
      security:
      - userToken: []
      summary: list events
      tags:
      - events
  /events/new:
    get:
      description: "<b>イベント情報取得</b><br>\n\t\t<ul>\n\t\t\t<li>/genre/:id -> ジャンル別新着情報</li>\n\t\t\t<li>/new
        -> ユーザー別新着情報</li>\n\t\t\t<li>/keep -> ユーザーがキープしているイベント</li>\n\t\t\t<li>/nokeep
        -> ユーザーが興味なしにしたイベント</li>\n\t\t\t<li>/popular -> キープ数が多い。注目されているイベント</li>\n\t\t\t<li>/recommend
        -> ユーザー属性に合わせたおすすめイベント</li>\n\t\t</ul>\n\t\tイベントの情報は区切って送信され、スクロールイベントで次のページのイベント情報を取得することを想定している。<br>\n\t\tまた、キープや興味なしの操作は１日に１回行われるバッチ処理時に確定されるまでは、分類されずに表示される。"
      operationId: events#list#1
      parameters:
      - in: query
        name: id
        required: false
        type: string
      - default: 0
        description: ページ(1->2->3->4)
        in: query
        minimum: 1
        name: page
        required: false
        type: integer
      - default: ""
        description: キーワード検索
        in: query
        name: q
        required: false
        type: string
      - default: ""
        description: ソート
        enum:
        - created_asc
        - created_desc
        - ""
        in: query
        name: sort
        required: false
        type: string
      produces:
      - application/vnd.goa.error
      - application/vnd.event+json; type=collection
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/EventCollection'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized
      schemes:
      - http
      security:
      - userToken: []
      summary: list events
      tags:
      - events
  /events/nokeep:
    get:
      description: "<b>イベント情報取得</b><br>\n\t\t<ul>\n\t\t\t<li>/genre/:id -> ジャンル別新着情報</li>\n\t\t\t<li>/new
        -> ユーザー別新着情報</li>\n\t\t\t<li>/keep -> ユーザーがキープしているイベント</li>\n\t\t\t<li>/nokeep
        -> ユーザーが興味なしにしたイベント</li>\n\t\t\t<li>/popular -> キープ数が多い。注目されているイベント</li>\n\t\t\t<li>/recommend
        -> ユーザー属性に合わせたおすすめイベント</li>\n\t\t</ul>\n\t\tイベントの情報は区切って送信され、スクロールイベントで次のページのイベント情報を取得することを想定している。<br>\n\t\tまた、キープや興味なしの操作は１日に１回行われるバッチ処理時に確定されるまでは、分類されずに表示される。"
      operationId: events#list#3
      parameters:
      - in: query
        name: id
        required: false
        type: string
      - default: 0
        description: ページ(1->2->3->4)
        in: query
        minimum: 1
        name: page
        required: false
        type: integer
      - default: ""
        description: キーワード検索
        in: query
        name: q
        required: false
        type: string
      - default: ""
        description: ソート
        enum:
        - created_asc
        - created_desc
        - ""
        in: query
        name: sort
        required: false
        type: string
      produces:
      - application/vnd.goa.error
      - application/vnd.event+json; type=collection
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/EventCollection'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized
      schemes:
      - http
      security:
      - userToken: []
      summary: list events
      tags:
      - events
  /events/popular:
    get:
      description: "<b>イベント情報取得</b><br>\n\t\t<ul>\n\t\t\t<li>/genre/:id -> ジャンル別新着情報</li>\n\t\t\t<li>/new
        -> ユーザー別新着情報</li>\n\t\t\t<li>/keep -> ユーザーがキープしているイベント</li>\n\t\t\t<li>/nokeep
        -> ユーザーが興味なしにしたイベント</li>\n\t\t\t<li>/popular -> キープ数が多い。注目されているイベント</li>\n\t\t\t<li>/recommend
        -> ユーザー属性に合わせたおすすめイベント</li>\n\t\t</ul>\n\t\tイベントの情報は区切って送信され、スクロールイベントで次のページのイベント情報を取得することを想定している。<br>\n\t\tまた、キープや興味なしの操作は１日に１回行われるバッチ処理時に確定されるまでは、分類されずに表示される。"
      operationId: events#list#4
      parameters:
      - in: query
        name: id
        required: false
        type: string
      - default: 0
        description: ページ(1->2->3->4)
        in: query
        minimum: 1
        name: page
        required: false
        type: integer
      - default: ""
        description: キーワード検索
        in: query
        name: q
        required: false
        type: string
      - default: ""
        description: ソート
        enum:
        - created_asc
        - created_desc
        - ""
        in: query
        name: sort
        required: false
        type: string
      produces:
      - application/vnd.goa.error
      - application/vnd.event+json; type=collection
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/EventCollection'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized
      schemes:
      - http
      security:
      - userToken: []
      summary: list events
      tags:
      - events
  /events/recommend:
    get:
      description: "<b>イベント情報取得</b><br>\n\t\t<ul>\n\t\t\t<li>/genre/:id -> ジャンル別新着情報</li>\n\t\t\t<li>/new
        -> ユーザー別新着情報</li>\n\t\t\t<li>/keep -> ユーザーがキープしているイベント</li>\n\t\t\t<li>/nokeep
        -> ユーザーが興味なしにしたイベント</li>\n\t\t\t<li>/popular -> キープ数が多い。注目されているイベント</li>\n\t\t\t<li>/recommend
        -> ユーザー属性に合わせたおすすめイベント</li>\n\t\t</ul>\n\t\tイベントの情報は区切って送信され、スクロールイベントで次のページのイベント情報を取得することを想定している。<br>\n\t\tまた、キープや興味なしの操作は１日に１回行われるバッチ処理時に確定されるまでは、分類されずに表示される。"
      operationId: events#list#5
      parameters:
      - in: query
        name: id
        required: false
        type: string
      - default: 0
        description: ページ(1->2->3->4)
        in: query
        minimum: 1
        name: page
        required: false
        type: integer
      - default: ""
        description: キーワード検索
        in: query
        name: q
        required: false
        type: string
      - default: ""
        description: ソート
        enum:
        - created_asc
        - created_desc
        - ""
        in: query
        name: sort
        required: false
        type: string
      produces:
      - application/vnd.event+json; type=collection
      - application/vnd.goa.error
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/EventCollection'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized
      schemes:
      - http
      security:
      - userToken: []
      summary: list events
      tags:
      - events
  /genres:
    get:
      description: "<b>ジャンル検索</b><br>\n\t\tジャンル名で検索し、当てはまるジャンルを返す。その際に対象となるジャンルがなかった場合、<br>\n\t\tジャンル追加ボタンを表示し、追加出来るようにする。"
      operationId: genres#list
      parameters:
      - default: 0
        description: ページ(1->2->3->4)
        in: query
        minimum: 1
        name: page
        required: false
        type: integer
      - default: ""
        description: ジャンル名検索に使うキーワード
        in: query
        maxLength: 30
        minLength: 0
        name: q
        required: false
        type: string
      - default: ""
        description: ソート
        enum:
        - created_asc
        - created_desc
        - ""
        in: query
        name: sort
        required: false
        type: string
      produces:
      - application/vnd.goa.error
      - application/vnd.genre+json; type=collection
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GenreCollection'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized
      schemes:
      - http
      security:
      - userToken: []
      summary: list genres
      tags:
      - genres
...
  /genres/new:
    post:
      description: "<b>ジャンルの新規作成</b><br>\n\t\t新しく作成するジャンル名を送信して、新規作成を行う。追加処理が完了とするとジャンルIDが返ってくるので、それを自動でフォローするようにする。<br>\n\t\t但し、ジャンルを新規作成する前に、ジャンル名を検索するフローを挟み、検索結果に出てこなかった場合に追加できるようにする。"
      operationId: genres#create
      parameters:
      - description: ジャンル名
        in: query
        maxLength: 30
        minLength: 1
        name: name
        required: true
        type: string
      produces:
      - application/vnd.goa.error
      - application/vnd.genre+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Genre'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized
      schemes:
      - http
      security:
      - userToken: []
      summary: create genres
      tags:
      - genres
...
produces:
- application/json
- application/xml
- application/gob
- application/x-gob
...
schemes:
- http
securityDefinitions:
  cronToken:
    in: header
    name: X-Appengine-Cron
    type: apiKey
  userToken:
    description: ユーザートークン
    in: header
    name: X-Authorization
    type: apiKey
swagger: "2.0"